Add .nojekill file
As stated in documentation:

GitHub Pages now uses GitHub Actions to execute the Jekyll build. When
using a branch as the source of your build, GitHub Actions must be
enabled in your repository if you want to use the built-in Jekyll
workflow. Alternatively, if GitHub Actions is unavailable or disabled,
adding a .nojekyll file to the root of your source branch will bypass
the Jekyll build process and deploy the content directly.